Feature: Add .env file and change main routine name
