--- src/index.js
@@ -1,65 +1,0 @@
-// You installed the `express` library earlier. For more information, see "[JavaScript example: Install dependencies](#javascript-example-install-dependencies)."
-const express = require('express')
-const bodyParser = require('body-parser')
-
-// This initializes a new Express application.
-const app = express()
-app.use(bodyParser.json())
-app.use(bodyParser.urlencoded({ extended: false }))
-
-// This defines a POST route at the `/webhook` path. This path matches the path that you specified for the smee.io forwarding. For more information, see "[Forward webhooks](#forward-webhooks)."
-//
-// Once you deploy your code to a server and update your webhook URL, you should change this to match the path portion of the URL for your webhook.
-app.post('/webhook', (request, response) => {
-  // Respond to indicate that the delivery was successfully received.
-  // Your server should respond with a 2XX response within 10 seconds of receiving a webhook delivery. If your server takes longer than that to respond, then GitHub terminates the connection and considers the delivery a failure.
-  response.status(202).send('Accepted')
-
-  // Check the `x-github-event` header to learn what event type was sent.
-  const githubEvent = request.headers['x-github-event']
-  console.log(`Incoming githubEvent: ${ githubEvent }`)
-
-  // You should add logic to handle each event type that your webhook is subscribed to.
-  // For example, this code handles the `issues` and `ping` events
-  //
-  // If any events have an `action` field, you should also add logic to handle each action that you are interested in.
-  // For example, this code handles the `opened` and `closed` actions for the `issue` event.
-  //
-  // For more information about the data that you can expect for each event type, see "[AUTOTITLE](/webhooks/webhook-events-and-payloads)."
-  if (githubEvent === 'issues') {
-    const body = JSON.parse(request.body.payload)
-    const action = body.action
-    console.log(`Invoked with action: ${ action } and data:`, body)
-    switch (action) {
-      case 'opened':
-        console.log(`An issue was opened with this title: ${ body.issue.title }`)
-        break
-      case 'assigned':
-        console.log(`A user was assigned to an issue ${ body.issue.assignee.login }`)
-        break
-      case 'labeled':
-        console.log(`A label was assigned to an issue ${ body.issue.label.name }`)
-        break
-      case 'closed':
-        console.log(`An issue was closed by ${ body.issue.user.login }`)
-        break
-      default:
-        console.log(`Unhandled action for the issue event: ${ action }`)
-    }
-  } else if (githubEvent === 'ping') {
-    console.log('GitHub sent the ping event')
-  } else {
-    console.log(`Unhandled event: ${githubEvent}`)
-  }
-});
-
-// This defines the port where your server should listen.
-// 3000 matches the port that you specified for webhook forwarding. For more information, see "[Forward webhooks](#forward-webhooks)."
-//
-// Once you deploy your code to a server, you should change this to match the port where your server is listening.
-const port = 3000
-
-// This starts the server and tells it to listen at the specified port.
-app.listen(port, () => {
-  console.log(`Server is running on port ${ port }`)
-})
